Szebben jelenik meg a félkész hozzáadás
