+++ frontend/anya-elmentem/src/views/addGroup/addGroup.vue
@@ -1,56 +1,89 @@
 <template>
-    <div class="addGroup">
-        <h1>Hozd létre a csoportodat!</h1>
-        <br>
-        <br>
-        <h2>Válaszd ki a csoportod a tagjait!</h2>
-            <select v-model="selected" multiple>
-              <option disabled value="" >Please select one</option>
-              <option v-for="i in users" v-bind:key='i' v-bind:value='i'  >{{i.email}}</option>
-            </select>
-            <br>
-        <form @submit="submit()" onsubmit = "return false">
-            <button type="submit">Csoport létrehozása</button>
+  <div class="addGroup-container">
+    <div class="row justify-content-center">
+      <b-jumbotron class="col-lg-4 col-md-6 col-sm-10">
+
+        <div class="col-12 title">
+          <h1>Új csoport</h1>
+          <hr class="my-4">
+        </div>
+
+        <div v-if="successfulNewGroup && !tried" class="col-12">
+          <h1>Csoprt létrehozás sikeres</h1>
+        </div>
+
+        <form v-if ="!successfulNewGroup || tried" class="row justify-content-center">
+          <multiselect
+          v-model="value"
+          :options="options"
+          :multiple="true"
+          :close-on-select="false"
+          :clear-on-select="false"
+          :preserve-search="true"
+          placeholder="Válaszd ki a csoport tagjait"
+          label="name"
+          track-by="name"
+          :preselect-first="true">
+          </multiselect>
+
+          <div class="col-md-5 col-sm-8 ">
+            <b-button variant="success" @click="submitGroup()">Létrehozás</b-button>
+          </div>
+
         </form>
+
+      </b-jumbotron>
     </div>
+  </div>
 </template>
 
 <script>
+import Multiselect from 'vue-multiselect'
 import { mapActions, mapState } from 'vuex'
+
 export default {
-  name: 'addGroup',
-  computed: {
-    ...mapState(['token']),
-    ...mapState('addGroup', ['users'])
-  },
-  methods: {
-    submit () {
-      this.addGroup({ token: this.token, memberList: this.selected })
-    },
-    ...mapActions('addGroup', ['addGroup', 'getUsers'])
+  components: {
+    Multiselect
   },
   data () {
     return {
-      selected: []
+      options: [],
+      value: [],
+      tried: true
     }
   },
   created () {
     this.getUsers({ token: this.token })
+  },
+  watch: {
+    users () {
+      this.options = this.users
+    }
+  },
+  computed: {
+    ...mapState(['token']),
+    ...mapState('addGroup', ['users', 'successfulNewGroup'])
+  },
+  methods: {
+    ...mapActions('addGroup', ['addGroupm', 'addGroup', 'getUsers']),
+    submitGroup () {
+      var emails = []
+      var i
+      for (i = 0; i < this.value.length; i++) {
+        emails.push(this.value[i].email)
+      }
+      this.addGroupm({ token: this.token, memberList: emails })
+      this.tried = false
+    }
   }
-
 }
 </script>
 
-<style scoped>
-h2{
-  color:red;
-}
-h1{
-  color:red;
-  text-align:left;
-}
+<style src="vue-multiselect/dist/vue-multiselect.min.css"></style>
 
-div{
-  background-color: black;
+<style scoped>
+.btn{
+  width: 100%;
+  margin-top: 15px;
 }
 </style>
